Move unique utility into gwpy.utils.misc
