+++ gwpy/utils/__init__.py
@@ -27,6 +27,7 @@ from .misc import (
     gprint,
     if_not_none,
     null_context,
+    unique,
 )
 
 __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
