+++ gwpy/utils/tests/test_misc.py
@@ -47,6 +47,13 @@ def test_null_context():
         print('this should work')
 
 
+def test_unique():
+    """Test for :func:`gwpy.utils.misc.unique`
+    """
+    a = [1, 2, 4, 3, 5, 4, 5, 3]
+    assert utils_misc.unique(a) == [1, 2, 4, 3, 5]
+
+
 @pytest.mark.parametrize('func, value, out', [
     (str, None, None),
     (str, 1, '1'),
